WIP for handling spaces in parameters
The previous code built a string using spaces as separators and then split it to provide the list of arguments. When input files contain spaces, this goes wrong.
The intent of this refactoring is to move the settings to QStringLists, so each part is intrinsically separated.

This commit is WIP; the encoder fails with an error at the moment.