WIP : fix cases where time remaining becomes 00:00:00
In some cases, the time remaining goes to 00:00:00. I see this fairly often for 10-bit encodes, or cases where the processing has been paused for some time and resumed. I think this may be related to an overflow or similar, so this is some playing around to try and understand/resolve it.
The current thought is to see how much time per frame is being spent and then use that with the remaining frames to predict the time to complete. By not considering the time-since-start, but the time-since-last-update, I'm hoping to make this more accurate and also more responsive to pause or system load variations.