Mixed bag commit
VobSub hardburn was successful. Now trying to work on the GUI. Split 'unsupported' helper to treat subtitles and audio separately and making the checks specific to the content type being reviewed. Making the title text indicate hard-burn support is required for 'unsupported' streams. The helpers don't yet consider the container, so this is configure more for MP4 output than MKV in terms of checking for validity.